countries_controller-refactor-to-efcore

+++ WebApplication452_simple/Controllers/CountriesController.cs
@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 using System.Collections.Generic;
 using System.Data;
 using System.Data.Entity;
@@ -17,21 +17,24 @@ namespace WebApplication452_simple.Controllers
         // GET: Countries
         public ActionResult Index()
         {
-            return View(db.Countries.ToList());
+            var countries = db.Countries.ToList();
+            return View(countries);
         }
 
         // GET: Countries/Details/5
         public ActionResult Details(int? id)
         {
-            if (id == null)
+            if (!id.HasValue)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
-            Country country = db.Countries.Find(id);
+
+            var country = db.Countries.Find(id);
             if (country == null)
             {
                 return HttpNotFound();
             }
+
             return View(country);
         }
 
@@ -42,8 +45,6 @@ namespace WebApplication452_simple.Controllers
         }
 
         // POST: Countries/Create
-        // To protect from overposting attacks, enable the specific properties you want to bind to, for 
-        // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
         public ActionResult Create([Bind(Include = "Id,Name")] Country country)
@@ -52,7 +53,7 @@ namespace WebApplication452_simple.Controllers
             {
                 db.Countries.Add(country);
                 db.SaveChanges();
-                return RedirectToAction("Index");
+                return RedirectToAction(nameof(Index));
             }
 
             return View(country);
@@ -61,21 +62,21 @@ namespace WebApplication452_simple.Controllers
         // GET: Countries/Edit/5
         public ActionResult Edit(int? id)
         {
-            if (id == null)
+            if (!id.HasValue)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
-            Country country = db.Countries.Find(id);
+
+            var country = db.Countries.Find(id);
             if (country == null)
             {
                 return HttpNotFound();
             }
+
             return View(country);
         }
 
         // POST: Countries/Edit/5
-        // To protect from overposting attacks, enable the specific properties you want to bind to, for 
-        // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
         public ActionResult Edit([Bind(Include = "Id,Name")] Country country)
@@ -84,23 +85,26 @@ namespace WebApplication452_simple.Controllers
             {
                 db.Entry(country).State = EntityState.Modified;
                 db.SaveChanges();
-                return RedirectToAction("Index");
+                return RedirectToAction(nameof(Index));
             }
+
             return View(country);
         }
 
         // GET: Countries/Delete/5
         public ActionResult Delete(int? id)
         {
-            if (id == null)
+            if (!id.HasValue)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
-            Country country = db.Countries.Find(id);
+
+            var country = db.Countries.Find(id);
             if (country == null)
             {
                 return HttpNotFound();
             }
+
             return View(country);
         }
 
@@ -109,10 +113,14 @@ namespace WebApplication452_simple.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
-            Country country = db.Countries.Find(id);
-            db.Countries.Remove(country);
-            db.SaveChanges();
-            return RedirectToAction("Index");
+            var country = db.Countries.Find(id);
+            if (country != null)
+            {
+                db.Countries.Remove(country);
+                db.SaveChanges();
+            }
+
+            return RedirectToAction(nameof(Index));
         }
 
         protected override void Dispose(bool disposing)
